config: rename base package
Renamed the base package to "at.aau.kuhhandel" in all modules and files.
