+++ app/src/androidTest/kotlin/at/aau/kuhhandel/androidapp/ExampleInstrumentedTest.kt
@@ -1,0 +1,16 @@
+package at.aau.kuhhandel.androidapp
+
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import androidx.test.platform.app.InstrumentationRegistry
+import org.junit.Assert.assertEquals
+import org.junit.Test
+import org.junit.runner.RunWith
+
+@RunWith(AndroidJUnit4::class)
+class ExampleInstrumentedTest {
+    @Test
+    fun usesCorrectPackageName() {
+        val appContext = InstrumentationRegistry.getInstrumentation().targetContext
+        assertEquals("at.aau.kuhhandel.androidapp", appContext.packageName)
+    }
+}
