--- app/src/androidTest/kotlin/com/example/androidapp/ExampleInstrumentedTest.kt
@@ -1,16 +1,0 @@
-package com.example.androidapp
-
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.platform.app.InstrumentationRegistry
-import org.junit.Assert.assertEquals
-import org.junit.Test
-import org.junit.runner.RunWith
-
-@RunWith(AndroidJUnit4::class)
-class ExampleInstrumentedTest {
-    @Test
-    fun usesCorrectPackageName() {
-        val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.androidapp", appContext.packageName)
-    }
-}
