+++ app/src/main/kotlin/at/aau/kuhhandel/androidapp/ui/theme/Theme.kt
@@ -1,0 +1,18 @@
+package at.aau.kuhhandel.androidapp.ui.theme
+
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.darkColorScheme
+import androidx.compose.material3.lightColorScheme
+import androidx.compose.runtime.Composable
+
+private val DarkColorScheme = darkColorScheme()
+private val LightColorScheme = lightColorScheme()
+
+@Composable
+fun AndroidAppTheme(darkTheme: Boolean = false, content: @Composable () -> Unit) {
+    val colorScheme = if (darkTheme) DarkColorScheme else LightColorScheme
+    MaterialTheme(
+        colorScheme = colorScheme,
+        content = content,
+    )
+}
