--- app/src/main/kotlin/com/example/androidapp/ui/menu/MainMenuScreen.kt
@@ -1,48 +1,0 @@
-package com.example.androidapp.ui.menu
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Button
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedButton
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-
-@Composable
-fun MainMenuScreen(modifier: Modifier = Modifier) {
-    Column(
-        modifier = modifier.fillMaxSize(),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center,
-    ) {
-        Text(
-            text = "Kuhhandel",
-            style = MaterialTheme.typography.headlineLarge,
-            modifier = Modifier.padding(bottom = 48.dp),
-        )
-
-        Button(
-            onClick = { /* TODO: Navigation zur Lobby/Spielersuche */ },
-            modifier = Modifier.fillMaxWidth(0.6f),
-        ) {
-            Text("Spiel starten")
-        }
-
-        Spacer(modifier = Modifier.height(16.dp))
-
-        OutlinedButton(
-            onClick = { /* TODO: Navigation zu den Regeln */ },
-            modifier = Modifier.fillMaxWidth(0.6f),
-        ) {
-            Text("Regeln")
-        }
-    }
-}
