--- app/src/main/kotlin/com/example/androidapp/ui/theme/Theme.kt
@@ -1,18 +1,0 @@
-package com.example.androidapp.ui.theme
-
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.darkColorScheme
-import androidx.compose.material3.lightColorScheme
-import androidx.compose.runtime.Composable
-
-private val DarkColorScheme = darkColorScheme()
-private val LightColorScheme = lightColorScheme()
-
-@Composable
-fun AndroidAppTheme(darkTheme: Boolean = false, content: @Composable () -> Unit) {
-    val colorScheme = if (darkTheme) DarkColorScheme else LightColorScheme
-    MaterialTheme(
-        colorScheme = colorScheme,
-        content = content,
-    )
-}
