+++ app/src/test/kotlin/at/aau/kuhhandel/androidapp/ExampleUnitTest.kt
@@ -1,0 +1,17 @@
+package at.aau.kuhhandel.androidapp
+
+import at.aau.kuhhandel.shared.Constants
+import kotlin.test.assertEquals
+import org.junit.Test
+
+class ExampleUnitTest {
+    @Test
+    fun `app version matches shared constant`() {
+        assertEquals("1.0.0", Constants.APP_VERSION)
+    }
+
+    @Test
+    fun `default server port is 8080`() {
+        assertEquals(8080, Constants.DEFAULT_PORT)
+    }
+}
