+++ server/src/main/kotlin/at/aau/kuhhandel/server/Application.kt
@@ -1,0 +1,36 @@
+package at.aau.kuhhandel.server
+
+import at.aau.kuhhandel.shared.ApiRoutes
+import at.aau.kuhhandel.shared.Constants
+import at.aau.kuhhandel.shared.HealthResponse
+import io.ktor.serialization.kotlinx.json.json
+import io.ktor.server.application.Application
+import io.ktor.server.application.call
+import io.ktor.server.application.install
+import io.ktor.server.engine.embeddedServer
+import io.ktor.server.netty.Netty
+import io.ktor.server.plugins.contentnegotiation.ContentNegotiation
+import io.ktor.server.response.respond
+import io.ktor.server.routing.get
+import io.ktor.server.routing.routing
+
+fun main() {
+    embeddedServer(Netty, port = Constants.DEFAULT_PORT) {
+        configurePlugins()
+        configureRoutes()
+    }.start(wait = true)
+}
+
+fun Application.configurePlugins() {
+    install(ContentNegotiation) {
+        json()
+    }
+}
+
+fun Application.configureRoutes() {
+    routing {
+        get(ApiRoutes.HEALTH) {
+            call.respond(HealthResponse(status = "UP"))
+        }
+    }
+}
