+++ shared/src/main/kotlin/at/aau/kuhhandel/shared/ApiRoutes.kt
@@ -1,0 +1,6 @@
+package at.aau.kuhhandel.shared
+
+/** Centralized API route constants used by both server and android-app. */
+object ApiRoutes {
+    const val HEALTH = "/health"
+}
