+++ shared/src/main/kotlin/at/aau/kuhhandel/shared/Constants.kt
@@ -1,0 +1,7 @@
+package at.aau.kuhhandel.shared
+
+/** App-wide constants shared across all modules. */
+object Constants {
+    const val APP_VERSION = "1.0.0"
+    const val DEFAULT_PORT = 8080
+}
