+++ shared/src/main/kotlin/at/aau/kuhhandel/shared/HealthResponse.kt
@@ -1,0 +1,13 @@
+package at.aau.kuhhandel.shared
+
+import kotlinx.serialization.Serializable
+
+/**
+ * Response model for the /health endpoint.
+ * Shared between android-app and server.
+ */
+@Serializable
+data class HealthResponse(
+    val status: String,
+    val message: String = "OK",
+)
