--- shared/src/main/kotlin/com/example/shared/ApiRoutes.kt
@@ -1,6 +1,0 @@
-package com.example.shared
-
-/** Centralized API route constants used by both server and android-app. */
-object ApiRoutes {
-    const val HEALTH = "/health"
-}
