--- shared/src/main/kotlin/com/example/shared/Constants.kt
@@ -1,7 +1,0 @@
-package com.example.shared
-
-/** App-wide constants shared across all modules. */
-object Constants {
-    const val APP_VERSION = "1.0.0"
-    const val DEFAULT_PORT = 8080
-}
